api/pyproject.toml
[tool.poetry]
name = "medicine-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "medicine_api"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
langchain = "^0.0.276"
fastapi = "^0.103.0"
uvicorn = "^0.23.2"
sqlalchemy = "^2.0.20"
python-dotenv = "^1.0.0"
pandas = "^2.1.0"
matplotlib = "^3.7.2"
scikit-learn = "^1.3.0"
keras = "^2.13.1"
tensorflow-io-gcs-filesystem = "0.31.0"
tensorflow = "^2.13.0"
prophet = "^1.1.4"
google-search-results = "^2.4.2"
beautifulsoup4 = "^4.12.2"
openai = "^0.28.0"
aioboto3 = "^11.3.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
